+++ queries/Query_1.js
@@ -1,0 +1,7 @@
+SELECT
+    name,
+    slug
+FROM
+    "prague-stories"
+WHERE
+    name LIKE CONCAT('%', :${请输入名称片段}, '%');