+++ queries/Query_2.js
@@ -1,0 +1,7 @@
+SELECT
+    name,
+    slug
+FROM
+    "locations" ,
+WHERE
+    name LIKE CONCAT('%', :${请输入名称片段}, '%');